+++ app/src/main/java/com/sevtinge/quickback/Prefs.java
@@ -7,5 +7,9 @@ public final class Prefs {
 
     public static final String FILE_NAME = "quick_back";
     public static final String KEY_ENABLED = "home_navigation_quick_back";
-}
+    public static final String KEY_SENSITIVITY = "home_navigation_quick_back_sensitivity";
 
+    public static final int SENSITIVITY_CONSERVATIVE = 0;
+    public static final int SENSITIVITY_STANDARD = 1;
+    public static final int SENSITIVITY_SENSITIVE = 2;
+}